Make sure Settings also work when running on Plasma
Moves the call into a C++ class.
Checks the system being run using the XDG_CURRENT_DESKTOP.
Prevents us from forwarding the QT_WAYLAND_SHELL_INTEGRATION value to
the child process, since it will not work as expected as we are
enforcing a private one. Made the switch for all desktops as I
understand it's the same problem elsewhere.