Allow disabling show/hide animations with an environment variable
Currently, the animation for showing or hiding the keyboard is done
client side. This leads to some nasty race conditions, at least on KWin
Wayland, because the keyboard rect is sent asynchronously and any window
geometry changes it causes are also async.

To allow the animation to be done by the compositor directly, we first
need to disable the client-side animation.