inputmethod: Revert to simpler style of word engine plugin loading
The new style of plugin loading seems to be failing for some of the plugins
despite the metadata being correct. This results in the plugin not being
loaded, and appropriate integrations not working. Reverting to the simple
check of whether the plugin.so exists results in the plugin always being
loaded. In the future, we should however get rid of these compiled plugin
libraries and extra QML, and use simple XML or JSON for the layout
definitions instead.