CMake: Restore Thai plug-in to not needing to link to westernsupport
Link the language plugins to the common library as they should have been,
and remove the abstractlanguageplugin.{cpp,h} from westernsupport as they
are already included in the keyboard library, linked by common. This allows
for languages using AbstractLanguagePlugin, but which don't need to link
against westersupport, such as Thai, to be built properly again.