CMake: Set OUTPUT_NAME on plugin libraries to the correct language name
Work around CMake's handling of @ inside argument names that then end up
getting defined as variable names in language_plugin as well. We also need
to set the OUTPUT_NAME of these plugins back to the unmodified language
value, so that the plugins can be properly loaded by the keyboard.