qml: Only load emoji keyboard QML when the state is EMOJI
With the change of emoji layout to being a state instead of a plugin, the
loader was always active, causing a slight reduction in performance with
showing the keyboard when requested. This change fixes that by only loading
the emoji QML when the emoji state is requested by the user.