Support using LayerShellQt for the input panel
inputpanel-shell is rather limited and doesn't support things like
proper placement and sizing. Fortunately, layer-shell can be used
instead to get support for these. This adds support for using
layer-shell, prefering it over inputpanel-shell. We only need to set
some properties from the Maliit side, most of the actual magic is
taken care of by the compositor.

Most importantly, this makes it possible for the compositor to ensure
Maliit does not overlap panels while at the same time ensuring nothing
ends up out of the screen.