qml: Implement a grid layout for longer extended keys lists
There are a couple of especially long extended key lists in some of the
keyboard layouts, so split them into two rows. This currently only supports
having two rows, but could be extended to up to three rows in the future
if necessary. Generally, keys shouldn't have so many extended keys on a
single character key, though.